@@ -28,8 +28,6 @@ public:
 	Eigen::VectorXd getdHdq();
 	void readStateCallback(const sensor_msgs::JointState::ConstPtr& msg);
 	Eigen::Vector3d getEEPose(Eigen::Matrix<double, 7, 1> q);
-	Eigen::Vector3d getCOG(int i, Eigen::Matrix<double, 7, 1> q);
-	Eigen::Matrix<double,4, 4> getTi(int i, Eigen::Matrix<double, 7, 1> q);
 	Eigen::VectorXd getdVdq();
 
 
@@ -47,7 +45,8 @@ private:
     Eigen::Matrix<double, 7, 1> DH_a, DH_d, DH_alpha;
     Eigen::Matrix<double, 3, 1> L_cogs[7];
 
-    const double mass[7] = {2.04471, 2.56414, 2.48050, 2.02754, 3.09611, 1.66736, 0.75606};
+    // The second mass should appear in dV dq (added 0)
+    const double mass[10] = {3.66357, 2.04471, 2.56414, 2.48050, 2.02754, 3.09611, 1.66736, 0.75606, 0.1};
 
 
     const double g = 9.81;
